requirements.txt
streamlit==1.28.1
pandas==2.1.1
plotly==5.17.0
openpyxl==3.1.2
xlrd==2.0.1
requests==2.31.0
python-dotenv==1.0.0
streamlit-authenticator==0.2.3
bcrypt==4.0.1
pyyaml==6.0.1
msal==1.25.0
openai==1.51.2
